+++ R/yarn_cluster.R
@@ -61,8 +61,61 @@ spark_yarn_cluster_get_app_property <- function(config, start_time, rm_webapp, p
   propertyValue
 }
 
+spark_yarn_cluster_resource_manager_is_online <- function(rm_webapp) {
+  rmQuery <- paste0(
+    "http",
+    "://",
+    rm_webapp,
+    "/ws/v1/cluster/info"
+  )
+
+  tryCatch({
+    !httr::http_error(rmQuery)
+  }, error = function(err) {
+    FALSE
+  })
+}
+
+spark_yarn_cluster_get_resource_manager_webapp <- function() {
+  rmHighAvailability <- spark_yarn_cluster_get_conf_property("yarn.resourcemanager.ha.enabled")
+  rmHighAvailability <- length(rmHighAvailability) > 0 && grepl("TRUE", rmHighAvailability, ignore.case = TRUE)
+
+  mainRMWebapp <- "yarn.resourcemanager.webapp.address"
+  if (rmHighAvailability) {
+    rmHighAvailabilityId <- spark_yarn_cluster_get_conf_property("yarn.resourcemanager.ha.id")
+
+    rmHighAvailabilityIds <- spark_yarn_cluster_get_conf_property("yarn.resourcemanager.ha.rm-ids")
+    rmHighAvailabilityIds <- strsplit(rmHighAvailabilityIds, ",")[[1]]
+    rmHighAvailabilityIds <- rmHighAvailabilityIds[rmHighAvailabilityIds != rmHighAvailabilityId]
+    rmHighAvailabilityIds <- c(rmHighAvailabilityId, rmHighAvailabilityIds)
+
+    mainRMWebapp <- NULL
+    for (rmId in rmHighAvailabilityIds) {
+      rmCandidate <- paste0("yarn.resourcemanager.webapp.address.", rmId)
+      rmCandidateValue <- spark_yarn_cluster_get_conf_property(rmCandidate)
+
+      if (spark_yarn_cluster_resource_manager_is_online(rmCandidateValue)) {
+        mainRMWebapp <- rmCandidate
+        break;
+      }
+    }
+
+    if (is.null(mainRMWebapp)) {
+      stop("Failed to find online resource manager under High Availability cluster.")
+    }
+  }
+
+  mainRMWebappValue <- spark_yarn_cluster_get_conf_property(mainRMWebapp)
+
+  if (is.null(mainRMWebappValue)) {
+    stop("Failed to retrieve ", mainRMWebapp, " from yarn-site.xml")
+  }
+
+  mainRMWebappValue
+}
+
 spark_yarn_cluster_get_gateway <- function(config, start_time) {
-  resourceManagerWebapp <- spark_yarn_cluster_get_conf_property("yarn.resourcemanager.webapp.address")
+  resourceManagerWebapp <- spark_yarn_cluster_get_resource_manager_webapp()
 
   if (length(resourceManagerWebapp) == 0) {
     stop("Yarn Cluster mode uses `yarn.resourcemanager.webapp.address` but is not present in yarn-site.xml")